return types, 64-bit changes: ioctl cmd -> u_long, casts, etc.
SCCS-vsn: 8.3

+++ sys/kern/tty_pty.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)tty_pty.c	8.2 (Berkeley) 09/23/93
+ *	@(#)tty_pty.c	8.3 (Berkeley) 01/09/95
  */
 
 /*
@@ -577,7 +577,7 @@ block:
 /*ARGSUSED*/
 ptyioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
